+++ backend/pipelined/handle.go
@@ -26,8 +26,8 @@ const (
 func (p *Pipelined) handleEvent(event *types.Event) error {
 	ctx := context.WithValue(context.Background(), types.OrganizationKey, event.Entity.Organization)
 	ctx = context.WithValue(ctx, types.EnvironmentKey, event.Entity.Environment)
-	handlers, err := p.expandHandlers(ctx, event.Check.Config.Handlers, 1)
 
+	handlers, err := p.expandHandlers(ctx, event.Check.Config.Handlers, 1)
 	if err != nil {
 		return err
 	}